Fix .with() chainable method for multiple arguments
Unless I'm wrong about the API, if I create a spy, call it like

        object.spy(5, 5, 3, 1, 8);

and then assert

        expect(object.spy).to.have.been.called.with(5, 5, 3, 1, 8)

it should pass. The problem is that the current code will compare each argument of the assertion with each argument of the actual call 5 times, and will fail with repeated arguments, because found > args.length. Looks like a bug to me, but I haven't seen examples of .with() with multiple arguments.